+++ Systems/RandomEncountersSystem.cs
@@ -116,7 +116,7 @@ namespace XPRising.Systems
 
             var user = Cache.SteamPlayerCache[steamID];
 
-            Output.SendLore(user.UserEntity, message);
+            Output.SendMessage(user.UserEntity, message);
             Plugin.Log(LoggingSystem, LogLevel.Info, $"Encounters started: {user.CharacterName} vs. {npcName}");
 
             if (RandomEncountersConfig.NotifyAdminsAboutEncountersAndRewards.Value)
@@ -124,7 +124,7 @@ namespace XPRising.Systems
                 var onlineAdmins = DataFactory.GetOnlineAdmins();
                 foreach (var onlineAdmin in onlineAdmins)
                 {
-                    Output.SendLore(onlineAdmin.UserEntity, $"Encounter started: {user.CharacterName} vs. {npcName}");
+                    Output.SendMessage(onlineAdmin.UserEntity, $"Encounter started: {user.CharacterName} vs. {npcName}");
                 }
             }
             RewardsMap[steamID][entity.Index] = DataFactory.GetRandomItem();
@@ -153,7 +153,7 @@ namespace XPRising.Systems
                         .Where(data => data.IsOnline && data.SteamID != userModel.PlatformId);
                     foreach (var player in onlineUsers)
                     {
-                        Output.SendLore(player.UserEntity, globalMessage);
+                        Output.SendMessage(player.UserEntity, globalMessage);
                     }
 
                 }
@@ -163,7 +163,7 @@ namespace XPRising.Systems
                         .Where(data => data.IsOnline && data.IsAdmin && data.SteamID != userModel.PlatformId);
                     foreach (var onlineAdmin in onlineAdmins)
                     {
-                        Output.SendLore(onlineAdmin.UserEntity, $"{userModel.CharacterName} earned an encounter reward: <color={itemModel.Color}>{itemModel.Name}</color>");
+                        Output.SendMessage(onlineAdmin.UserEntity, $"{userModel.CharacterName} earned an encounter reward: <color={itemModel.Color}>{itemModel.Name}</color>");
                     }
                 }
             }